Deduplicate the avro/json comparison impls
